move clock stuff to file
