+++ src/Ilúvatar/iluvatar_load_gen/src/benchmark.rs
@@ -5,9 +5,8 @@ use clap::Parser;
 use iluvatar_controller_library::server::controller_comm::ControllerAPIFactory;
 use iluvatar_library::types::{CommunicationMethod, Compute, Isolation, MemSizeMb, ResourceTimings};
 use iluvatar_library::utils::config::args_to_json;
-use iluvatar_library::{logging::LocalTime, transaction::gen_tid, utils::port_utils::Port};
+use iluvatar_library::{clock::LocalTime, transaction::gen_tid, utils::port_utils::Port};
 use serde::{Deserialize, Serialize};
-use std::sync::Arc;
 use std::time::{Duration, SystemTime};
 use std::{collections::HashMap, path::Path};
 use iluvatar_library::tokio_utils::{build_tokio_runtime, TokioRuntime};
@@ -144,7 +143,7 @@ pub async fn benchmark_controller(
     for function in &functions {
         let mut func_data = FunctionStore::new(function.image_name.clone(), function.name.clone());
         println!("{}", function.name);
-        let clock = LocalTime::new(&gen_tid())?;
+        let clock = LocalTime::boxed(&gen_tid())?;
         let reg_tid = gen_tid();
         let api = factory
             .get_controller_api(&host, port, CommunicationMethod::RPC, &reg_tid)
@@ -223,7 +222,7 @@ pub fn benchmark_worker(threaded_rt: &TokioRuntime, functions: Vec<ToBenchmarkFu
     }
     let mut invokes = vec![];
     let factory = iluvatar_worker_library::worker_api::worker_comm::WorkerAPIFactory::boxed();
-    let clock = LocalTime::new(&gen_tid())?;
+    let clock = LocalTime::boxed(&gen_tid())?;
     let mut cold_repeats = args.cold_iters;
 
     for function in &functions {
