+++ src/Ilúvatar/iluvatar_worker_library/src/services/invocation/cpu_q_invoke.rs
@@ -15,7 +15,7 @@ use crate::worker_api::worker_config::{FunctionLimits, InvocationConfig};
 use anyhow::Result;
 use iluvatar_library::characteristics_map::CharacteristicsMap;
 use iluvatar_library::{
-    logging::LocalTime, threading::tokio_runtime, threading::EventualItem, transaction::TransactionId, types::Compute,
+    clock::LocalTime, threading::tokio_runtime, threading::EventualItem, transaction::TransactionId, types::Compute,
 };
 use parking_lot::Mutex;
 use std::{
@@ -25,6 +25,7 @@ use std::{
 use time::OffsetDateTime;
 use tokio::sync::{mpsc::UnboundedSender, Notify};
 use tracing::{debug, error, info, warn};
+use iluvatar_library::clock::GlobalClock;
 
 lazy_static::lazy_static! {
   pub static ref INVOKER_CPU_QUEUE_WORKER_TID: TransactionId = "InvokerCPUQueue".to_string();
