+++ src/Ilúvatar/iluvatar_worker_library/src/services/invocation/gpu_q_invoke.rs
@@ -22,7 +22,7 @@ use crate::{
 use anyhow::Result;
 use iluvatar_library::{characteristics_map::CharacteristicsMap, types::DroppableToken};
 use iluvatar_library::{
-    logging::LocalTime, threading::tokio_runtime, threading::EventualItem, transaction::TransactionId, types::Compute,
+    clock::LocalTime, threading::tokio_runtime, threading::EventualItem, transaction::TransactionId, types::Compute,
 };
 use parking_lot::Mutex;
 use std::collections::VecDeque;
@@ -33,6 +33,7 @@ use std::{
 use time::OffsetDateTime;
 use tokio::sync::Notify;
 use tracing::{debug, error, info, warn};
+use iluvatar_library::clock::GlobalClock;
 
 lazy_static::lazy_static! {
   pub static ref INVOKER_GPU_QUEUE_WORKER_TID: TransactionId = "InvokerGPUQueue".to_string();
@@ -486,7 +487,6 @@ impl DeviceQueue for GpuQueueingInvoker {
 #[cfg(test)]
 mod gpu_batch_tests {
     use super::*;
-    use iluvatar_library::logging::LocalTime;
     use std::collections::HashMap;
 
     fn item(clock: &LocalTime) -> Arc<EnqueuedInvocation> {
