+++ src/Ilúvatar/iluvatar_worker_library/src/services/invocation/queueing/gpu_mqfq.rs
@@ -12,7 +12,7 @@ use anyhow::Result;
 use dashmap::{mapref::multiple::RefMutMulti, DashMap};
 use iluvatar_library::types::{Compute, DroppableToken};
 use iluvatar_library::utils::missing_default;
-use iluvatar_library::{characteristics_map::CharacteristicsMap, logging::LocalTime, transaction::TransactionId};
+use iluvatar_library::{characteristics_map::CharacteristicsMap, clock::LocalTime, transaction::TransactionId};
 use iluvatar_library::{
     mindicator::Mindicator,
     threading::{tokio_runtime, tokio_thread, EventualItem},
@@ -26,6 +26,7 @@ use std::time::{Duration, Instant};
 use time::OffsetDateTime;
 use tokio::sync::Notify;
 use tracing::{debug, error, info, warn};
+use iluvatar_library::clock::GlobalClock;
 
 lazy_static::lazy_static! {
   pub static ref MQFQ_GPU_QUEUE_WORKER_TID: TransactionId = "MQFQ_GPU_Queue".to_string();
