+++ src/Ilúvatar/iluvatar_worker_library/src/services/invocation/queueing_dispatcher.rs
@@ -13,12 +13,13 @@ use crate::worker_api::worker_config::{FunctionLimits, GPUResourceConfig, Invoca
 use anyhow::Result;
 use iluvatar_library::characteristics_map::CharacteristicsMap;
 use iluvatar_library::types::ComputeEnum;
-use iluvatar_library::{logging::LocalTime, transaction::TransactionId, types::Compute};
+use iluvatar_library::{clock::LocalTime, transaction::TransactionId, types::Compute};
 use parking_lot::RwLock;
 use rand::Rng;
 use std::{collections::HashMap, sync::Arc};
 use time::OffsetDateTime;
 use tracing::{debug, info};
+use iluvatar_library::clock::GlobalClock;
 
 lazy_static::lazy_static! {
   pub static ref INVOKER_CPU_QUEUE_WORKER_TID: TransactionId = "InvokerCPUQueue".to_string();
