+++ src/Ilúvatar/iluvatar_worker_library/tests/mqfq_tests.rs
@@ -5,7 +5,7 @@ use crate::utils::{short_sim_args, sim_args, sim_invoker_svc};
 use iluvatar_library::mindicator::Mindicator;
 use iluvatar_library::types::{Compute, Isolation};
 use iluvatar_library::{
-    logging::LocalTime,
+    clock::LocalTime,
     transaction::{gen_tid, TEST_TID},
 };
 use iluvatar_rpc::rpc::{LanguageRuntime, RegisterRequest};
@@ -23,6 +23,7 @@ use rstest::rstest;
 use std::collections::HashMap;
 use std::sync::Arc;
 use time::{Duration, OffsetDateTime};
+use iluvatar_library::clock::GlobalClock;
 
 static TIMEOUT_SEC: f64 = 2.0;
